Fix authentication: in some case, the user was not recognized
When the user already exists into the database, the corresponding
record was not given to Jelix, and so to other modules that
access to the user object had wrong information. In some case, for
example the jCommunity module, the status field was then empty
so jCommunity considered that the user is not allowed to connect.